requirements.txt
pyttsx3
SpeechRecognition
wikipedia
pywhatkit
pyautogui
keyboard
googlesearch-python
requests
python-dotenv
